requirements.txt
Django==4.1
djangorestframework==3.14.0
python-dotenv==1.0.0
django-cors-headers==3.13.0
PyJWT==2.6.0
pytz==2022.6
